refactor(newsclient): Move overriding method to private
This is so we can set sane defaults without having to copy the defaults
for every implementation.  It also moves the wrapping of articles and
the enforcement of limiting to be more general.

Additionally, add a limit variable.  It is enforced at the base-level,
but sent to the overridden method so that it may choose to optimize with
it.